# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Guardian API",
    "version": "1.0.0",
    "description": "API documentation for the Guardian Monitoring App"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullname",
                  "email",
                  "password",
                  "role"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Registration unsuccessful. Could be due to missing fields or invalid email/password."
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Log in a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login with JWT token and user information."
          },
          "400": {
            "description": "Bad request. Invalid credentials or account locked."
          }
        }
      }
    },
    "/api/v1/auth/send-pin": {
      "post": {
        "summary": "Send OTP for email verification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent to your email address."
          },
          "400": {
            "description": "Bad request. Email is required."
          },
          "500": {
            "description": "Internal server error. Error processing the request."
          }
        }
      }
    },
    "/api/v1/auth/verify-pin": {
      "post": {
        "summary": "Verifies OTP sent to user email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "otp"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "otp": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully."
          },
          "400": {
            "description": "Bad request. Email and OTP are required or invalid/expired OTP."
          },
          "500": {
            "description": "Internal server error. Error processing the request."
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "summary": "Change a user's password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "oldPassword",
                  "newPassword",
                  "confirmPassword"
                ],
                "properties": {
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully."
          },
          "400": {
            "description": "Bad request. Incorrect old password or new passwords don't match."
          }
        }
      }
    },
    "/api/v1/auth/reset-password-request": {
      "post": {
        "summary": "Request a password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent to the user's email."
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "get": {
        "summary": "Render password reset page",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "JWT reset token"
          }
        ],
        "responses": {
          "200": {
            "description": "Renders password reset page."
          },
          "400": {
            "description": "Invalid or expired token."
          }
        }
      },
      "post": {
        "summary": "Reset's a user's password using the reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "newPassword",
                  "confirmPassword"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "JWT_TOKEN_HERE"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "NewPassword123"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "NewPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password has been updated successfully."
          },
          "400": {
            "description": "Invalid or expired token, or passwords don't match."
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}